Liv-CSS
Hi, this is the CSS I wrote for my index page. It has a bit of the navigation bar formatting and other things we mainly just did in class! I found a page of colour palettes developed by W3Schools.com (https://www.w3schools.com/colors/colors_palettes.asp), and I thought that might be a useful way to go about designing our colour scheme! At the top of the script there's the /*"#?????"*/ which are the different colours.