CSS edits
I've just played around with the CSS here - nothing much worth noting here (as in nothing that we can't change) except for the ".recipleimage", which is the class I've assigned to the images of the cocktails used for the cocktail images on the recipe pages so that we can format them in the same size, etc. throughout the whole site.