CSS - new design
Here's the new design I've been working on. I've added functionality like a drop-down menu in the navigation bar and boxes for the content to be contained in which add to the general aesthetic and organisation of the page. The colour scheme, fonts etc. can be changed and experimented with. The colour palettes are from the W3Schools CSS website.